FIX: Lecture 상세정보에 찜한 개수 추가

+++ devoot-backend/src/main/java/com/gamee/devoot_backend/lecture/dto/LectureDetail.java
@@ -11,10 +11,10 @@ public record LectureDetail(
 	String sourceUrl,
 	String imgUrl,
 	String curriculum,
-	int bookmarkCount,
+	long bookmarkCount,
 	float rating
 ) {
-	public LectureDetail(Lecture lecture, int bookmarkCount, float rating) {
+	public LectureDetail(Lecture lecture, long bookmarkCount, float rating) {
 		this(
 			lecture.getName(),
 			lecture.getLecturer(),
